requirements.txt
streamlit==1.29.0
fastapi==0.109.0
uvicorn==0.24.0
requests==2.31.0
openai==1.3.0
pydantic==2.5.3
nltk>=3.8.1,<4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
